% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file with FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name of the file to read. This contains the data}
}
\value{
This returns a dataframe after reading the csv file. If no file is present then it
returns an error
}
\description{
This function takes in \strong{US National Highway Traffic Safety Administration's}
data from .csv file, stored on disk. This data contains a nationwide census, providing the
American public yearly data, regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\examples{
fars_read('accident_2013.csv.bz2')

}
