% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Input list of years \code{years}}
}
\value{
This function returns a data frame which includes data for a given list of years \code{years}
}
\description{
This function takes an input as \code{years} and then filters the information for that given list
}
\examples{
fars_read_years(2014)

}
