% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Input list of years \code{years}}
}
\value{
This function returns a data frame which includes data for a given list of years \code{years}
It summarizes the information by counting number of accidents happening in a given month, year.
}
\description{
This function takes an input as \code{years} and then summarizes the information for that given list
}
\examples{
fars_summarize_years(2013)

}
