% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Makes a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Input a year \code{year} in form of character or integer}
}
\value{
This function returns a file in CSV format which is related to the given \code{year}
}
\description{
This function makes a file in CSV format which is related to the given \code{year}
}
\examples{
make_filename(2013)

}
